Update pal from commit: 4dbde31
* Set AMDGPU_IB_FLAG_EMIT_MEM_SYNC flag
* Add support for TMZ(trusted memory zone) feature
* Update ilInstructionSet.cpp with new IL opcodes, and attempt to update existing opcodes with correct data for 'noModifiers'
* Fix NGG small prim filter culling with custom sample positions
* Fix app hang / crash related to shader ring change
* Fix  Trackmania 2 glitchy graphical issue
* Make pal separately reads Pal settings files for different PAL clients. For Vulkan, Pal will read amdVulkanSettings.cfg. If amdVulkanSettings.cfg does not exist, read amdPalSettings,cfg
* Fix assertion failure triggering in  PAL + >=GFX9 when importing memory allocation into a 2D D32_S8 image
* Fix vkGetRandROutputDisplayEXT on repeated calls
* Query perf experiment Properties
* Fix PAL_ASSERT bug in Pal::Amdgpu::Device::CreateGpuMemoryFromExternalShare
* Fix typo for QueryProfiling memory clock
* Correct a few errors in PAL's subgroup size calculation for NGG GS
* Pack DepthStencilStateCreateInfo better to reduce the size
* Remove the numScratchWavesPerCu public setting
* [cmake]Create pal_include_guard helper function
* Bump version number to 255

+++ inc/core/pal.h
@@ -213,7 +213,7 @@ enum GpuHeap : uint32
 
 /// Comparison function determines how a pass/fail condition is determined between two values.  For depth/stencil
 /// comparison, the first value comes from source data and the second value comes from destination data.
-enum class CompareFunc : uint32
+enum class CompareFunc : uint8
 {
     Never        = 0x0,
     Less         = 0x1,
